chore: Add @mercadopago/sdk-react dependency and update Firebase SDK version
This commit adds the "@mercadopago/sdk-react" dependency to the project, with version "^0.0.19". It also updates the Firebase SDK version to "10.13.0" and initializes Firestore in the project. The necessary environment variables for Firebase have been added to the `.env` file. The Firebase configuration has been set up in the `firebasesdk.js` file, using the provided environment variables.

Closes #123